default to float32 rotate/randrotate

+++ tests/test_inverse.py
@@ -212,7 +212,12 @@ TESTS.append(("RandZoom 3d", "3D", 9e-2, RandZoomd(KEYS, 1, [0.5, 0.6, 0.9], [1.
 TESTS.append(("RandRotated, prob 0", "2D", 0, RandRotated(KEYS, prob=0)))
 
 TESTS.append(
-    ("Rotated 2d", "2D", 8e-2, Rotated(KEYS, random.uniform(np.pi / 6, np.pi), keep_size=True, align_corners=False))
+    (
+        "Rotated 2d",
+        "2D",
+        8e-2,
+        Rotated(KEYS, random.uniform(np.pi / 6, np.pi), keep_size=True, align_corners=False, dtype=np.float64),
+    )
 )
 
 TESTS.append(
@@ -220,7 +225,7 @@ TESTS.append(
         "Rotated 3d",
         "3D",
         1e-1,
-        Rotated(KEYS, [random.uniform(np.pi / 6, np.pi) for _ in range(3)], True),  # type: ignore
+        Rotated(KEYS, [random.uniform(np.pi / 6, np.pi) for _ in range(3)], True, dtype=np.float64),  # type: ignore
     )
 )
